pyproject.toml
[tool.poetry]
name = "jsonify-resume"
version = "0.1.1"
description = "A cli that converts resumes into JSON Resume schema"
authors = ["Ashish Binu <ashishbinu90@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "jsonify_resume" }]
include = ["prompt", "schema.json"]

[tool.poetry.scripts]
jsonify-resume = "jsonify_resume.main:app"

[tool.poetry.dependencies]
python = "^3.11"
revchatgpt = "^6.8.6"
python-magic = "^0.4.27"
pdftotext = "^2.2.2"
python-dotenv = "^1.0.0"
typer = { extras = ["all"], version = "^0.9.0" }
rich = "^13.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
